// src/Ui.Angular/src/app/shared/generated/model/include-active-option.ts
/**
 * Kaizen.Skills.Service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: latest
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type IncludeActiveOption = 'None' | 'IncludeOnlyActive' | 'IncludeOnlyInactive' | 'IncludeAll';

export const IncludeActiveOption = {
    None: 'None' as IncludeActiveOption,
    IncludeOnlyActive: 'IncludeOnlyActive' as IncludeActiveOption,
    IncludeOnlyInactive: 'IncludeOnlyInactive' as IncludeActiveOption,
    IncludeAll: 'IncludeAll' as IncludeActiveOption
};