// src/Ui.Angular/src/app/shared/generated/model/skill-tree-item.ts
/**
 * Kaizen.Skills.Service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: latest
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SkillAggregationLevel } from './skill-aggregation-level';

export interface SkillTreeItem { 
    Id?: number;
    Name?: string;
    IsActive?: boolean;
    NodeType?: SkillAggregationLevel;
    Items?: Array<SkillTreeItem>;
}